added route '/user' to manage users

+++ network_storage_state.go
@@ -5,8 +5,8 @@ type NetworkStorageState struct {
 	NumberOfMonthlySubs     int64  `bson:"number_of_monthly_subs"`
 	NumberOfFixedAmount1Subs int64  `bson:"number_of_fixed_amount1_subs"`
 	NumberOfFixedAmount2Subs int64  `bson:"number_of_fixed_amount2_subs"`
-	TotalAwsStorageSize     int64  `bson:"total_aws_storage_size"`  // in gigabytes
-	TotalAwsStorageUsed     int64  `bson:"total_aws_storage_used"`  // in gigabytes
-	TotalStoragePoolSize    int64  `bson:"total_storage_pool_size"` // in gigabytes
-	TotalStoragePoolUsed    int64  `bson:"total_storage_pool_used"` // in gigabytes
+	TotalAwsStorageSize     float64  `bson:"total_aws_storage_size"`  // in gigabytes
+	TotalAwsStorageUsed     float64  `bson:"total_aws_storage_used"`  // in gigabytes
+	TotalStoragePoolSize    float64  `bson:"total_storage_pool_size"` // in gigabytes
+	TotalStoragePoolUsed    float64  `bson:"total_storage_pool_used"` // in gigabytes
 }
